Criacao de servicos e controllers

+++ src/main/java/com/tcd/repository/ConteudoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tcd.model.Conteudo;
 
-public interface ConteudoRepository extends JpaRepository<Conteudo, Integer> {
+public interface ConteudoRepository extends JpaRepository<Conteudo, Long> {
 
 }
